[linux-port] Implement thread-safe incr/decr (#201)
InterlockedIncrement and InterlockedDecrement are atomic operations to
increment or decrement the value pointed to by the pointer parameter and
return the same. Currently they are implemented as simple addition
operations in WinMacros.h for non-win32 platforms which works well
enough, but wasn't really the original intent.

There are already platform abstractions available to LLVM in Atomics.h.
This change leverages them to implement the same functionality in a
platform-agnostic way. Mostly it's just a matter of swapping
InterlockedIncrement for AtomicIncrement.

Fixes half of https://github.com/google/DirectXShaderCompiler/issues/198