[linux-port] Fix static function header warnings (#219)
Having fully implemented functions in a header file means that
every source file that includes it will get its own set of these
functions whether it uses them or not. If they are unused, they
are reported in an error as a defined but unused function. In a
common header, this happens a lot. By leaving only prototypes in
the header and putting the bodies in a source file, they are all
silenced.
Fixes 363 clang and 363 gcc warnings

Contributes to https://github.com/google/DirectXShaderCompiler/issues/206