Create utility IsComputeLike functions
For both shadermodel and functionproperties objects, we frequently
wanted to know if a shader stage was "compute-like" which this defines
to mean can take numthreads, but deals with aspects besides numthreads
specifically.

Adding these revealed a number of cases where we were checking the wrong
things and will need tests for.

signatures should be on all "graphics" stages except amplification and
no others

numthreads can be in CS/MS/AS/Node

viewID should be in ???

Rename some functions that inaccurately state they are identifying
whether there should be signatures or not

remove isgraphicsshader and iscomputeshader from shadermodel class
because all it did was extract the props from the passed function and
use that to get the answer and everywhere it was called, the properties
were already ready available