Finalish restructure of test for readability
Establish explicit 32-bit or 64-bit versions of types otherwise dependent on the RUN line that get set whenever is appropriate for the given RUN. This clarifies what a captured variable means at any given time. This required using -DAG to process the same line twice in some cases to capture the same variable as RUN dependent and explicitly sized, but it really improved readability and allowed for consolidation of most CHECK lines.