this is probably a bad idea
I did this to try to sus out whether the code that applied unsigned qualifiers to shorthand vector types needed to be changed. It allows for long vecs in signatures, but I don't think we want that. Still a useful tool to identify those areas in case I need to revisit. Revert imminent