modify longvecs to accept many-sized vecs
Not sure I want this to be the same test. Might want to parcel it out so I can test vec8 thoroughly in other respects and test variable sizes independently. Anyway, nothing much came of this, everything meaningful is fine.

There were some weird mangles I had to deal with and I couldn't test the vector load and creation as thoroughly in a size-robust way, so I just focused on the final touch of the vector creation which gave me the register I needed to ensure the ops got created properly.