Adapt raw buffer structs to common load lowering
When raw buffers use a templated load with a struct, they reuse the
subscript path also used for subscripted structured buffers. Such loads
with structs containing vectors or matrices will invoke the load
lowering from within this recursive call that traverses GEPs and other
users of the original call to set up correct offsets etc.

This adapts that code to use the common load lowering that enables long
vectors within structs to be correctly loaded.

Since the code expects byte address buffers, it is not (yet) adapted to
structured buffers, so those code paths are kept as they were.