Refactor resloadhelper constructors
the default offset behavior for non-call instructions was different in every location so making it explicit and letting each call location calculate it as appropriate cleaned things up. It turns out the only case of matrix loading where the type was thought to be different from the type of the value to replace was always the same, so the type member is removed. The only case where offset calculation was different for the call instruction constructor was also where the replaced instruction needed to be explicit, so the boolean parameter was replaced by the explicit replacement instruction which sets that boolean and explicitly sets the replaced instruction to either the matrix load instruction or a load instruction of the result of a subscript operator.