[Cooperative Vectors] Require use of unwritable input types
For matrix and vector ref inputs, the use of the *Impl template allowed passing in writable RWMatrixRef and RWVectorRef. This wasn't desired and ultimately failed with a confusing error on the builtin functions. By changing the types of the intrinsics, read-only types are required. This has the consequence of removing RWVectorRef entirely and changing some of the error messages to match the new signatures by changing the intrinsic header implementations to take the specific types, which requires defining them individually.

Fixes #7490